refactor(web): abstract out collapse and rotate animation logic into their own components
