+++ apps/web/src/components/Cards/SurfaceCards/BlobTransactionCard.tsx
@@ -1,11 +1,12 @@
-import React, { useCallback, useEffect, useRef, useState } from "react";
+import React, { useState } from "react";
 import type { FC } from "react";
 import { ArrowRightIcon, ChevronDownIcon } from "@heroicons/react/24/outline";
-import { animated, useSpring } from "@react-spring/web";
 
 import { Button } from "~/components/Button";
+import { Collapsable } from "~/components/Collapsable";
 import { EtherUnitDisplay } from "~/components/Displays/EtherUnitDisplay";
 import { RollupIcon } from "~/components/RollupIcon";
+import { Rotable } from "~/components/Rotable";
 import { Skeleton } from "~/components/Skeleton";
 import { useBreakpoint } from "~/hooks/useBreakpoint";
 import {
@@ -21,29 +22,6 @@ import { RollupBadge } from "../../Badges/RollupBadge";
 import { Link } from "../../Link";
 import { CardField } from "../Card";
 import { SurfaceCardBase } from "./SurfaceCardBase";
-
-const CollapseIcon: React.FC<{
-  opened: boolean;
-  onClick(): void;
-}> = function ({ onClick, opened }) {
-  const props = useSpring({
-    from: { rotate: 0 },
-    to: { rotate: Number(opened) * 180 },
-  });
-
-  return (
-    // eslint-disable-next-line jsx-a11y/click-events-have-key-events, jsx-a11y/no-static-element-interactions
-    <div className="-p cursor-pointer" onClick={onClick}>
-      <animated.div style={props} className="-mb-2">
-        <Button
-          variant="icon"
-          icon={<ChevronDownIcon className="h-5 w-5" />}
-          size="md"
-        />
-      </animated.div>
-    </div>
-  );
-};
 
 type BlobTransactionCardProps = Partial<{
   transaction: Partial<
@@ -93,12 +71,6 @@ const BlobTransactionCard: FC<BlobTransactionCardProps> = function ({
   compact,
 }) {
   const [opened, setOpened] = useState(false);
-  const contentRef = useRef<HTMLDivElement | null>(null);
-  const contentHeight = useRef<number>(0);
-  const props = useSpring({
-    from: { openProgress: 0 },
-    to: { openProgress: Number(opened) },
-  });
   const breakpoint = useBreakpoint();
   const isCompact =
     compact ||
@@ -106,25 +78,9 @@ const BlobTransactionCard: FC<BlobTransactionCardProps> = function ({
     breakpoint === "md" ||
     breakpoint === "default";
   const displayBlobs = !compact && !!blobsOnTx?.length;
-
-  const updateHeight = useCallback(() => {
-    if (contentRef.current) {
-      contentHeight.current = contentRef.current.clientHeight;
-    }
-  }, []);
-
-  const handleContentRef = useCallback(
-    (element: HTMLDivElement) => {
-      contentRef.current = element;
-      updateHeight();
-    },
-    [updateHeight]
-  );
-
-  useEffect(updateHeight, [opened, updateHeight]);
-
   const totalBlobSize = blobsOnTx?.reduce((acc, { size }) => acc + size, 0);
 
+  console.log(displayBlobs);
   return (
     <div>
       <SurfaceCardBase
@@ -234,29 +190,25 @@ const BlobTransactionCard: FC<BlobTransactionCardProps> = function ({
           </div>
           {displayBlobs && (
             <div className="-mb-2 flex items-center justify-center md:-mt-5">
-              <CollapseIcon
-                opened={opened}
-                onClick={() => {
-                  setOpened((op) => !op);
-                }}
-              />
+              <Rotable
+                angle={180}
+                rotated={opened}
+                onClick={() => setOpened((prevOpened) => !prevOpened)}
+              >
+                <Button
+                  variant="icon"
+                  icon={<ChevronDownIcon className="h-5 w-5" />}
+                  size="md"
+                />
+              </Rotable>
             </div>
           )}
         </div>
       </SurfaceCardBase>
-      {displayBlobs && (
-        <div className="overflow-hidden bg-primary-200 pr-4 dark:bg-primary-900">
-          <animated.div
-            style={{
-              height: props.openProgress.to(
-                (value) => `${value * contentHeight.current}px`
-              ),
-            }}
-          >
-            <div
-              ref={handleContentRef}
-              className="ml-10 grid grid-cols-[1fr_6fr_2fr] gap-2 p-2 text-sm"
-            >
+      {blobsOnTx && (
+        <Collapsable opened={opened}>
+          <div className="bg-primary-200 pr-4 dark:bg-primary-900">
+            <div className="ml-10 grid grid-cols-[1fr_6fr_2fr] gap-2 p-2 text-sm">
               <TableHeader>Index</TableHeader>
               <TableHeader>Versioned Hash</TableHeader>
               <TableHeader>Size</TableHeader>
@@ -272,8 +224,8 @@ const BlobTransactionCard: FC<BlobTransactionCardProps> = function ({
                 </React.Fragment>
               ))}
             </div>
-          </animated.div>
-        </div>
+          </div>
+        </Collapsable>
       )}
     </div>
   );
